+++ Modules/Intent.Modules.Eventing.AzureEventGrid/Templates/IntegrationEventHandler/IntegrationEventHandlerTemplatePartial.cs
@@ -6,6 +6,7 @@ using Intent.Modules.Common;
 using Intent.Modules.Common.CSharp.Builder;
 using Intent.Modules.Common.CSharp.Templates;
 using Intent.Modules.Common.Templates;
+using Intent.Modules.Eventing.Shared.Templates;
 using Intent.RoslynWeaver.Attributes;
 using Intent.Templates;
 
@@ -14,40 +15,18 @@ using Intent.Templates;
 
 namespace Intent.Modules.Eventing.AzureEventGrid.Templates.IntegrationEventHandler
 {
-    [IntentManaged(Mode.Fully, Body = Mode.Merge)]
-    public partial class IntegrationEventHandlerTemplate : CSharpTemplateBase<IntegrationEventHandlerModel>, ICSharpFileBuilderTemplate
+    [IntentManaged(Mode.Ignore)]
+    public partial class IntegrationEventHandlerTemplate : IntegrationEventHandlerTemplateBase
     {
+        [IntentManaged(Mode.Fully)]
         public const string TemplateId = "Intent.Eventing.AzureEventGrid.IntegrationEventHandler";
 
-        [IntentManaged(Mode.Fully, Body = Mode.Ignore)]
-        public IntegrationEventHandlerTemplate(IOutputTarget outputTarget, IntegrationEventHandlerModel model) : base(TemplateId, outputTarget, model)
+        public IntegrationEventHandlerTemplate(IOutputTarget outputTarget, IntegrationEventHandlerModel model)
+            : base(
+                templateId: TemplateId,
+                outputTarget: outputTarget,
+                model: model)
         {
-            CSharpFile = new CSharpFile(this.GetNamespace(), this.GetFolderPath())
-                .AddClass($"{Model.Name}", @class =>
-                {
-                    @class.AddConstructor(ctor =>
-                    {
-                        ctor.AddParameter("string", "exampleParam", param =>
-                        {
-                            param.IntroduceReadonlyField();
-                        });
-                    });
-                });
-        }
-
-        [IntentManaged(Mode.Fully)]
-        public CSharpFile CSharpFile { get; }
-
-        [IntentManaged(Mode.Fully)]
-        protected override CSharpFileConfig DefineFileConfig()
-        {
-            return CSharpFile.GetConfig();
-        }
-
-        [IntentManaged(Mode.Fully)]
-        public override string TransformText()
-        {
-            return CSharpFile.ToString();
         }
     }
 }