+++ Modules/Intent.Modules.Eventing.AzureEventGrid/Templates/PublisherOptions/PublisherOptionsTemplate.cs
@@ -12,14 +12,27 @@ namespace Intent.Modules.Eventing.AzureEventGrid.Templates.PublisherOptions
         {
             return $$"""
                      using System;
-
+                     using System.Collections.Generic;
+                     
                      [assembly: DefaultIntentManaged(Mode.Fully)]
-
+                     
                      namespace {{Namespace}}
                      {
                          public class {{ClassName}}
                          {
+                             private readonly List<PublisherEntry> _entries = [];
+                             
+                             public IReadOnlyList<PublisherEntry> Entries => _entries;
+                             
+                             public void Add<TMessage>(string credentialKey, string endpoint)
+                             {
+                                 ArgumentNullException.ThrowIfNull(credentialKey);
+                                 ArgumentNullException.ThrowIfNull(endpoint);
+                                 _entries.Add(new PublisherEntry(typeof(TMessage), credentialKey, endpoint));
+                             }
                          }
+                         
+                         public record PublisherEntry(Type MessageType, string CredentialKey, string Endpoint);
                      }
                      """;
         }