+++ Modules/Intent.Modules.Eventing.AzureEventGrid/Templates/SubscriptionOptions/SubscriptionOptionsTemplate.cs
@@ -1,3 +1,4 @@
+using Intent.Modules.Eventing.Contracts.Templates;
 using Intent.RoslynWeaver.Attributes;
 
 [assembly: DefaultIntentManaged(Mode.Fully)]
@@ -12,14 +13,32 @@ namespace Intent.Modules.Eventing.AzureEventGrid.Templates.SubscriptionOptions
         {
             return $$"""
                      using System;
-
+                     using System.Collections.Generic;
+                     using System.Threading;
+                     using System.Threading.Tasks;
+                     using Azure.Messaging.ServiceBus;
+                     
                      [assembly: DefaultIntentManaged(Mode.Fully)]
-
+                     
                      namespace {{Namespace}}
                      {
                          public class {{ClassName}}
                          {
+                             private readonly List<SubscriptionEntry> _entries = [];
+                             
+                             public IReadOnlyList<SubscriptionEntry> Entries => _entries;
+                             
+                             public void Add<TMessage, THandler>()
+                                 where TMessage : class
+                                 where THandler : {{this.GetIntegrationEventHandlerInterfaceName()}}<TMessage>
+                             {
+                                 _entries.Add(new SubscriptionEntry(typeof(TMessage), {{this.GetAzureEventGridMessageDispatcherName()}}.InvokeDispatchHandler<TMessage, THandler>));
+                             }
                          }
+                         
+                         public delegate Task DispatchHandler(IServiceProvider serviceProvider, ServiceBusReceivedMessage message, CancellationToken cancellationToken);
+                         
+                         public record SubscriptionEntry(Type MessageType, DispatchHandler HandlerAsync);
                      }
                      """;
         }