+++ Modules/Intent.Modules.Eventing.AzureEventGrid/release-notes.md
@@ -1,1 +1,3 @@
 ### Version 1.0.0
+
+- New Feature: Model Topics in Intent Architect to generate code that directly integrates with Azure Event Grid.